requirements.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
ants==0.0.7
antspyx==0.4.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==6.1.0
blessed==1.19.1
cachetools==5.2.0
cffi==1.15.1
cfgv==3.4.0
chart-studio==1.1.0
ConfigArgParse==1.5.3
connected-components-3d==3.12.4
contourpy==1.2.0
coverage==6.5.0
crc32c==2.3
cycler==0.11.0
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
degrade @ git+https://gitlab.com/iacl/degrade@c2755a7df4783c85ab676016301cada47131f183
dicom2nifti==2.4.10
dill==0.3.8
dipy==1.5.0
distlib==0.3.8
Django==4.1
elasticdeform==0.5.0
ema-pytorch==0.0.9
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.9.1
fastjsonschema==2.16.1
fastremap==1.14.1
filelock==3.13.1
fill-voids==2.0.6
Flask==2.2.2
flyingcircus==0.1.4.0
flyingcircus-numeric==0.1.1.2
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.4.0
func-timeout==4.3.5
gevent==21.12.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grad-cam==1.4.6
greenlet==1.1.2
grpcio==1.47.0
h5py==3.7.0
identify==2.5.34
imageio==2.34.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
llvmlite==0.42.0
Markdown==3.4.1
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.3
mistune==0.8.4
monai==1.3.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==3.2.1
nibabel==5.2.0
ninja==1.10.2.4
nodeenv==1.8.0
notebook==6.4.12
numba==0.59.0
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.2.3
open3d==0.16.0
opencv-python==4.6.0.66
openpyxl==3.1.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
piq==0.7.0
platformdirs==4.2.0
plotly==5.19.0
pooch==1.8.1
pre-commit==3.6.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pydicom==2.4.4
Pygments==2.17.2
pyparsing==3.1.1
pyquaternion==0.9.9
pyrsistent==0.18.1
pytest==7.4.4
python-dateutil==2.8.2
python-gdcm==3.0.23
pytorch-lightning==2.2.0.post0
pytz==2024.1
pyvista==0.43.3
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
raster-geometry==0.1.4.1
requests-oauthlib==1.3.1
resize @ git+https://gitlab.com/iacl/resize@7edd72d4a14d0b650e90313ab63f9d690a189ced
retrying==1.3.4
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
scooby==0.9.2
Send2Trash==1.8.0
setuptools-scm==7.0.5
sigpy==0.1.26
SimpleITK==2.3.1
simplification==0.7.10
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
statsmodels==0.14.1
sympy==1.12
taming-transformers==0.0.1
tenacity==8.2.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
terminado==0.15.0
tfrecord==1.14.1
threadpoolctl==3.3.0
tifffile==2024.2.12
tinycss2==1.1.1
tk==0.1.0
tomli==2.0.1
torch==2.2.1+cu118
torch-fidelity==0.3.0
torchaudio==2.2.1+cu118
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.17.1+cu118
tornado==6.2
TPTBox==0.0.5
tqdm==4.66.2
traitlets==5.3.0
transforms3d==0.4.1
triton==2.2.0
ttach==0.0.3
typing_extensions==4.9.0
tzdata==2024.1
virtualenv==20.25.0
vtk==9.3.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
yarl==1.9.4
zope.event==4.5.0
zope.interface==5.4.0
